플레이어 대쉬 및 디테일 작업
220922 / 오후 11시 43분

+++ Assets/01.Script/1.HyunWoong/Interface/IDamage.cs
@@ -8,5 +8,5 @@ public interface IDamage
     /// 요소들이 받는 피해
     /// </summary>
     /// <param name="damage">받을 피해</param>
-    public abstract void OnDamaged(float damage);
+    public abstract void OnDamaged(int damage);
 }
